refactor: Move some helpers to common code
To make sure no code in src/api or src/common depends on any code in src/server or src/client
