+++ src/common/helpers/appCleanup.js
